+++ StripeFinancialConnections/StripeFinancialConnections/Source/API Bindings/Models/FinancialConnectionsAuthSession.swift
@@ -76,18 +76,26 @@ struct FinancialConnectionsAuthSession: Decodable {
     let skipAccountSelection: Bool?
     let url: String?
     let isOauth: Bool?
+    let display: Display?
 
     var isOauthNonOptional: Bool {
         return isOauth ?? false
     }
 
     var requiresNativeRedirect: Bool {
-        guard flow == .mxOauthAppToApp else { return false }
         return url?.hasNativeRedirectPrefix ?? false
     }
 
     var partner: FinancialConnectionsPartner? {
         return (showPartnerDisclosure ?? false) ? flow?.toPartner() : nil
+    }
+
+    struct Display: Decodable {
+        let text: Text?
+
+        struct Text: Decodable {
+            let oauthPrepane: FinancialConnectionsOAuthPrepane?
+        }
     }
 }