+++ StripeFinancialConnections/StripeFinancialConnections/Source/API Bindings/Models/FinancialConnectionsSessionManifest.swift
@@ -45,6 +45,12 @@ struct FinancialConnectionsSessionManifest: Decodable {
         case unparsable
     }
 
+    enum ManualEntryMode: String, SafeEnumCodable, Equatable {
+        case automatic
+        case custom
+        case unparsable
+    }
+
     // MARK: - Properties
 
     let accountholderIsLinkConsumer: Bool?
@@ -77,4 +83,6 @@ struct FinancialConnectionsSessionManifest: Decodable {
     let features: [String: Bool]?
     let experimentAssignments: [String: String]?
     let assignmentEventId: String?
+    let skipSuccessPane: Bool?
+    let manualEntryMode: ManualEntryMode
 }