+++ StripeFinancialConnections/StripeFinancialConnections/Source/Native/AccountPicker/LinkingAccountsLoadingView.swift
@@ -36,14 +36,14 @@ final class LinkingAccountsLoadingView: UIView {
                     if let businessName = businessName {
                         return String(
                             format: STPLocalizedString(
-                                "Please wait while your account is linked to %@.",
+                                "Please wait while your account is connected to %@.",
                                 "The subtitle/description of the loading screen that appears when a user is in process of connecting their bank account to an application. Once the bank account is connected (or linked), the user will be able to use the bank account for payments.  %@ will be replaced by the business name, for example, The Coca-Cola Company."
                             ),
                             businessName
                         )
                     } else {
                         return STPLocalizedString(
-                            "Please wait while your account is linked to Stripe.",
+                            "Please wait while your account is connected to Stripe.",
                             "The subtitle/description of the loading screen that appears when a user is in process of connecting their bank account to an application. Once the bank account is connected (or linked), the user will be able to use the bank account for payments."
                         )
                     }
@@ -51,14 +51,14 @@ final class LinkingAccountsLoadingView: UIView {
                     if let businessName = businessName {
                         return String(
                             format: STPLocalizedString(
-                                "Please wait while your accounts are linked to %@.",
+                                "Please wait while your accounts are connected to %@.",
                                 "The subtitle/description of the loading screen that appears when a user is in process of connecting their bank accounts to an application. Once the bank accounts are connected (or linked), the user will be able to use those bank accounts for payments.  %@ will be replaced by the business name, for example, The Coca-Cola Company."
                             ),
                             businessName
                         )
                     } else {
                         return STPLocalizedString(
-                            "Please wait while your accounts are linked to Stripe.",
+                            "Please wait while your accounts are connected to Stripe.",
                             "The subtitle/description of the loading screen that appears when a user is in process of connecting their bank accounts to an application. Once the bank accounts are connected (or linked), the user will be able to use those bank accounts for payments."
                         )
                     }