+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ### PaymentSheet
 * [Added] Adds support for setting up PayPal using a SetupIntent or a PaymentIntent w/ setup_future_usage=off_session. Note: PayPal is in beta.
 
+## 23.3.4 2023-02-13
+### Financial Connections
+* [Changed] Polished Financial Connections UI.
+
 ## 23.3.3 2023-01-30
 ### Payments
 * [Changed] Updated image asset for AFFIN bank.